Added -Wall flag to gcc, and fixed a mountain of warnings ;-)

+++ HTSearchParser.cpp
@@ -106,8 +106,8 @@ HTSearchParser::_ParseNodes(BList* nodeList, BList* resultList)
 			status = B_ERROR;
 		else {
 			//Parse (Format: "martinhpedersen (Martin H. Pedersen)"
-			int screenNameEndIndex = buffer.find(" (");
-			int fullNameEndIndex = buffer.find(")");
+			size_t screenNameEndIndex = buffer.find(" (");
+			size_t fullNameEndIndex = buffer.find(")");
 			
 			if(fullNameEndIndex != std::string::npos) {
 				currentTweet->setFullName(buffer.substr(screenNameEndIndex+2, fullNameEndIndex-2-screenNameEndIndex).c_str());
@@ -136,10 +136,10 @@ HTSearchParser::_ParseNodes(BList* nodeList, BList* resultList)
 			status = B_ERROR;
 		else {
 			// Parse the data for Application name
-            int pos = buffer.find(">", 0); //<a href="http://www.tweetdeck.com/" rel="nofollow">TweetDeck</a>
+            size_t pos = buffer.find(">", 0); //<a href="http://www.tweetdeck.com/" rel="nofollow">TweetDeck</a>
 			if(pos != std::string::npos) {
-				int start = pos;
-				int end = pos;
+				size_t start = pos;
+				size_t end = pos;
 				while(end < buffer.length() && buffer[end] != '<') {
 					end++;
 				}
